scim-config-file-timestamp: Changed format to store timestamps
In the past revision, the timestamp of configuration file was stored as
a single integer which was transformed from a pair of integers, but its
value was too large so it was out of range and might cause unexpected
behavior. Therefore we modified the program in order to store the
timestamps as it is.